rename to cr-dev-server
