+++ cqf-fhir-cr-dev-server/src/main/java/org/opencds/cqf/fhir/cr/dev/server/config/CqlOperationConfig.java
@@ -1,0 +1,31 @@
+package org.opencds.cqf.fhir.cr.dev.server.config;
+
+import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.context.FhirVersionEnum;
+import ca.uhn.fhir.rest.server.RestfulServer;
+import java.util.List;
+import java.util.Map;
+import org.opencds.cqf.fhir.cr.hapi.common.ICqlProcessorFactory;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderLoader;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderSelector;
+import org.opencds.cqf.fhir.cr.hapi.r4.cql.CqlExecutionOperationProvider;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class CqlOperationConfig {
+    @Bean
+    CqlExecutionOperationProvider r4CqlExecutionOperationProvider(ICqlProcessorFactory cqlProcessorFactory) {
+        return new CqlExecutionOperationProvider(cqlProcessorFactory);
+    }
+
+    @Bean(name = "cqlOperationLoader")
+    public ProviderLoader cqlOperationLoader(
+            ApplicationContext applicationContext, FhirContext fhirContext, RestfulServer restfulServer) {
+        var selector = new ProviderSelector(
+                fhirContext, Map.of(FhirVersionEnum.R4, List.of(CqlExecutionOperationProvider.class)));
+
+        return new ProviderLoader(restfulServer, applicationContext, selector);
+    }
+}
