+++ cqf-fhir-cr-dev-server/src/main/java/org/opencds/cqf/fhir/cr/dev/server/config/EvaluateOperationConfig.java
@@ -1,0 +1,31 @@
+package org.opencds.cqf.fhir.cr.dev.server.config;
+
+import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.context.FhirVersionEnum;
+import ca.uhn.fhir.rest.server.RestfulServer;
+import java.util.Arrays;
+import java.util.Map;
+import org.opencds.cqf.fhir.cr.hapi.common.ILibraryProcessorFactory;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderLoader;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderSelector;
+import org.opencds.cqf.fhir.cr.hapi.r4.library.LibraryEvaluateProvider;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class EvaluateOperationConfig {
+    @Bean
+    LibraryEvaluateProvider r4LibraryEvaluateProvider(ILibraryProcessorFactory libraryProcessorFactory) {
+        return new LibraryEvaluateProvider(libraryProcessorFactory);
+    }
+
+    @Bean(name = "evaluateOperationLoader")
+    public ProviderLoader evaluateOperationLoader(
+            ApplicationContext applicationContext, FhirContext fhirContext, RestfulServer restfulServer) {
+        var selector = new ProviderSelector(
+                fhirContext, Map.of(FhirVersionEnum.R4, Arrays.asList(LibraryEvaluateProvider.class)));
+
+        return new ProviderLoader(restfulServer, applicationContext, selector);
+    }
+}
