+++ cqf-fhir-cr-dev-server/src/main/java/org/opencds/cqf/fhir/cr/dev/server/config/PopulateOperationConfig.java
@@ -1,0 +1,32 @@
+package org.opencds.cqf.fhir.cr.dev.server.config;
+
+import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.context.FhirVersionEnum;
+import ca.uhn.fhir.rest.server.RestfulServer;
+import java.util.Arrays;
+import java.util.Map;
+import org.opencds.cqf.fhir.cr.hapi.common.IQuestionnaireProcessorFactory;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderLoader;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderSelector;
+import org.opencds.cqf.fhir.cr.hapi.r4.questionnaire.QuestionnairePopulateProvider;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class PopulateOperationConfig {
+    @Bean
+    QuestionnairePopulateProvider r4QuestionnairePopulateProvider(
+            IQuestionnaireProcessorFactory questionnaireProcessorFactory) {
+        return new QuestionnairePopulateProvider(questionnaireProcessorFactory);
+    }
+
+    @Bean(name = "populateOperationLoader")
+    public ProviderLoader populateOperationLoader(
+            ApplicationContext applicationContext, FhirContext fhirContext, RestfulServer restfulServer) {
+        var selector = new ProviderSelector(
+                fhirContext, Map.of(FhirVersionEnum.R4, Arrays.asList(QuestionnairePopulateProvider.class)));
+
+        return new ProviderLoader(restfulServer, applicationContext, selector);
+    }
+}
