+++ cqf-fhir-cr-dev-server/src/main/java/org/opencds/cqf/fhir/cr/dev/server/config/QuestionnaireOperationConfig.java
@@ -1,0 +1,32 @@
+package org.opencds.cqf.fhir.cr.dev.server.config;
+
+import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.context.FhirVersionEnum;
+import ca.uhn.fhir.rest.server.RestfulServer;
+import java.util.Arrays;
+import java.util.Map;
+import org.opencds.cqf.fhir.cr.hapi.common.IQuestionnaireProcessorFactory;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderLoader;
+import org.opencds.cqf.fhir.cr.hapi.config.ProviderSelector;
+import org.opencds.cqf.fhir.cr.hapi.r4.structuredefinition.StructureDefinitionQuestionnaireProvider;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class QuestionnaireOperationConfig {
+    @Bean
+    StructureDefinitionQuestionnaireProvider r4StructureDefinitionQuestionnaireProvider(
+            IQuestionnaireProcessorFactory questionnaireProcessorFactory) {
+        return new StructureDefinitionQuestionnaireProvider(questionnaireProcessorFactory);
+    }
+
+    @Bean(name = "questionnaireOperationLoader")
+    public ProviderLoader questionnaireOperationLoader(
+            ApplicationContext applicationContext, FhirContext fhirContext, RestfulServer restfulServer) {
+        var selector = new ProviderSelector(
+                fhirContext, Map.of(FhirVersionEnum.R4, Arrays.asList(StructureDefinitionQuestionnaireProvider.class)));
+
+        return new ProviderLoader(restfulServer, applicationContext, selector);
+    }
+}
