--- cqf-fhir-cr-server/src/main/java/org/opencds/cqf/fhir/cr/server/config/ExtractOperationConfig.java
@@ -1,32 +1,0 @@
-package org.opencds.cqf.fhir.cr.server.config;
-
-import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.context.FhirVersionEnum;
-import ca.uhn.fhir.rest.server.RestfulServer;
-import java.util.Arrays;
-import java.util.Map;
-import org.opencds.cqf.fhir.cr.hapi.common.IQuestionnaireResponseProcessorFactory;
-import org.opencds.cqf.fhir.cr.hapi.config.ProviderLoader;
-import org.opencds.cqf.fhir.cr.hapi.config.ProviderSelector;
-import org.opencds.cqf.fhir.cr.hapi.r4.questionnaireresponse.QuestionnaireResponseExtractProvider;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ExtractOperationConfig {
-    @Bean
-    QuestionnaireResponseExtractProvider r4QuestionnaireResponseExtractProvider(
-            IQuestionnaireResponseProcessorFactory questionnaireResponseProcessorFactory) {
-        return new QuestionnaireResponseExtractProvider(questionnaireResponseProcessorFactory);
-    }
-
-    @Bean(name = "extractOperationLoader")
-    public ProviderLoader extractOperationLoader(
-            ApplicationContext applicationContext, FhirContext fhirContext, RestfulServer restfulServer) {
-        var selector = new ProviderSelector(
-                fhirContext, Map.of(FhirVersionEnum.R4, Arrays.asList(QuestionnaireResponseExtractProvider.class)));
-
-        return new ProviderLoader(restfulServer, applicationContext, selector);
-    }
-}
