+++ settings.gradle.kts
@@ -29,7 +29,7 @@ include("cqf-fhir-cr-spring")
 
 include("cqf-fhir-cr-cli")
 
-include("cqf-fhir-cr-server")
+include("cqf-fhir-cr-dev-server")
 
 include("cqf-fhir-benchmark")
